Simply put - huge rewrite.
I've changed the way how fluffy works so now it supports some neat stuff.

Fox example: users are now able to request the deps they need when the class
is created, they can also register and listen to events with parameters
and much more.

There's even more to come and this commit is by no means a working version.
It's more like a sketch of what's to come